Updating and Removing links
I've removed the Inknownledge link, it appears that API has gone away.
I've changed Bestparking link to Arrive. ParkWiz purchased Bestparking then it looks like ParkWiz is rebranding to Arrive. Which is why the API has the parkwhiz domain.
I've updated a bunch of links to point to where ever the old URL redirected you to.